merge develop in ft8

+++ lib/Drivers/STM32H7xx_HAL_Driver/Inc/stm32h7xx_hal_hcd.h
@@ -241,7 +241,7 @@ HAL_StatusTypeDef HAL_HCD_UnRegisterHC_NotifyURBChangeCallback(HCD_HandleTypeDef
 HAL_StatusTypeDef HAL_HCD_HC_SubmitRequest(HCD_HandleTypeDef *hhcd, uint8_t ch_num,
                                            uint8_t direction, uint8_t ep_type,
                                            uint8_t token, uint8_t *pbuff,
-                                           uint16_t length, uint8_t do_ping);
+										   uint32_t length, uint8_t do_ping);
 
 /* Non-Blocking mode: Interrupt */
 void HAL_HCD_IRQHandler(HCD_HandleTypeDef *hhcd);
